[embed] Rename SignOperationModal to OperationModal
